DCE-103: Decouple & Refactor Move Ordering
