+++ src/engine/move_generator/mod.rs
@@ -1,3 +1,4 @@
 pub mod make_move;
 pub mod mv_gen;
+pub mod mv_oredering;
 pub mod perft;
