+++ src/lib.rs
@@ -15,6 +15,7 @@ pub mod engine {
     pub mod move_generator {
         pub mod make_move;
         pub mod mv_gen;
+        pub mod mv_oredering;
         pub mod perft;
     }
     pub mod attacks {
